add logic splitting for fetching matches to service
